tests/submission/W3C/domhtml/HTMLLinkElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>W3C</title>
<link href="/resources/testharness.css" hreflang="en" media="screen" rev="stylesheet" type="text/css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-link-element.html#dom-link-disabled" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The disabled attribute enables/disables the link.   

    Retrieve the disabled attribute and examine its value.</p>

<script>
var domtest = async_test("The disabled attribute enables/disables the link.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vdisabled;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("link");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(0);
      vdisabled = testNode.disabled;

      assert_false(vdisabled);

    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
